+++ src/main/java/com/dta/services/dao/IUserDao.java
@@ -6,7 +6,7 @@ import com.dta.services.model.User;
 
 public interface IUserDao {
 	void create(User user);
-	User getById(int id);
+	User getById(long id);
 	List<User> list();
 	void update(User user);
 	void delete(int id);
